requirements.txt
Flask
requests
python-dotenv
itsdangerous
Werkzeug
